fpgainfo: refactor code for print phy information
Move the common code for printing phy information from board_n6000.c
to board_common.c.
