+++ libraries/libboard/board_common/CMakeLists.txt
@@ -28,5 +28,5 @@ opae_add_static_library(TARGET board_common
     SOURCE
         board_common.c
         ${opae-test_ROOT}/framework/mock/opae_std.c
-    LIBS opae-c
+    LIBS opae-c opaeuio
 )
