+++ tests/fpgainfo/CMakeLists.txt
@@ -53,10 +53,12 @@ opae_test_add(TARGET test_fpgainfo_c
     SOURCE test_fpgainfo_c.cpp
     LIBS fpgainfo-static
          board-common-static
+         opaeuio
 )
 
 opae_test_add(TARGET test_fpgainfo_board_c 
     SOURCE test_board_c.cpp
     LIBS fpgainfo-static
          board-common-static
+         opaeuio
 )
